requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiosqlite==0.17.0
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
clikit==0.6.2
cmdstanpy==0.9.68
convertdate==2.3.2
crashtest==0.3.1
cx-Oracle==8.2.1
cycler==0.10.0
Cython==0.29.24
databases==0.4.3
distlib==0.3.2
ephem==4.0.0.2
et-xmlfile==1.1.0
fastapi==0.68.0
filelock==3.0.12
hijri-converter==2.2.0
holidays==0.11.2
httpstan==4.5.0
idna==3.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lz4==3.1.3
marshmallow==3.13.0
matplotlib==3.4.3
multidict==5.1.0
mysql-connector-python==8.0.26
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
pastel==0.2.1
Pillow==8.3.1
platformdirs==2.2.0
prophet==1.0.1
protobuf==3.17.3
psycopg2-binary==2.9.1
pydantic==1.8.2
pylev==1.4.0
PyMeeus==0.5.11
PyMySQL==0.9.3
pyodbc==4.0.32
pyparsing==2.4.7
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
setuptools-git==1.2
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
tqdm==4.62.2
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
virtualenv==20.7.1
webargs==7.0.1
yarl==1.6.3
